Actualizacion

+++ Assets/Scripts/GameManager.cs
@@ -5,12 +5,12 @@ using UnityEngine;
 public class GameManager : MonoBehaviour
 {
     public int enemigosDerrotados;
-    public int vida;
+    public float vida;
     // Start is called before the first frame update
     void Start()
     {
         enemigosDerrotados = 0;
-        vida = 100;
+        vida = 100f;
     }
 
     // Update is called once per frame
